Re-read the web-id token for each S3 request
The token file pointed to by the AWS_WEB_IDENTITY_TOKEN_FILE env var
was read once at startup, but for long operations the token might
expire before we're complete. This switches to reading it on each
S3 request so we immediately use the renewed token. This is especially
likely when running in kubernetes using a projected service account
token.

Fixes #2428